modules/location-api/producer/requirements.txt
Flask
kafka-python
gunicorn
flask-accepts==0.10.0
flask-cors==3.0.8
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-Script==2.0.6
